+++ src/edu/hitsz/bullet/BaseBullet.java
@@ -13,7 +13,7 @@ public class BaseBullet extends AbstractFlyingObject {
 
     private int power = 10;
 
-    public BaseBullet(int locationX, int locationY, int speedX, int speedY, int power) {
+    public BaseBullet(double locationX, double locationY, double speedX, double speedY, int power) {
         super(locationX, locationY, speedX, speedY);
         this.power = power;
     }
