+++ src/edu/hitsz/supply/HpSupply.java
@@ -5,7 +5,7 @@ import edu.hitsz.application.WorldHandle;
 public class HpSupply extends BaseSupply {
     private int increase;
 
-    public HpSupply(int x, int y, int speedX, int speedY, int increase) {
+    public HpSupply(double x, double y, double speedX, double speedY, int increase) {
         super(x, y, speedX, speedY);
         this.increase = increase;
     }
