+++ src/edu/hitsz/supply/HpSupplyFactory.java
@@ -7,7 +7,7 @@ public class HpSupplyFactory implements SupplyFactory{
         return INSTANCE;
     }
     @Override
-    public BaseSupply createSupply(int x, int y, int vx, int vy) {
+    public BaseSupply createSupply(double x, double y, double vx, double vy) {
         return new HpSupply(x, y, vx, vy, 20);
     }
 }
